ml_training/requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
librosa==0.10.2
scikit-learn==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
tqdm==4.66.4
tensorboard==2.17.0
pandas==2.2.2
scipy==1.13.0
imagehash==4.3.1
faiss-cpu==1.8.0
